// src/types/openapi_MS.ts
export interface components {
    schemas: {
// ...
        ResGradeDTO: {
            /** Format: int64 */
            id?: number;
            name?: string;
            /** Format: int64 */
            studentsInPeriodCount?: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}